backend/reaxAcademi/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.15
gunicorn==25.1.0
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.12.0
